config: extract address-book compilation to compiler_security_addressbook.go
Part B of #4121 (pure code motion). Moves the address-book cluster
verbatim out of compiler_security.go, including the zone-local
address-book resolution (#3061) and the operator-facing display
helpers:

  - zoneLocalNamePrefix, zoneLocalQualify, ZoneLocalUnqualify
  - DisplayAddressName / DisplayAddressNames
  - resolveZoneLocalAddressBooks (the #3061/#3287 zone-qualified fold)
  - compileAddressBook / parseAddressBookEntries
  - mergeAddressNode / descriptionText / looksLikeIPOrCIDR

The zone-local helpers sit near the top of the original file while the
global-book parsers sit mid-file; their relative order is preserved.
parseAddressBookEntries is shared with compileZones (now in
compiler_security_zones.go) — both stay in package config, so the
cross-file call resolves with no import change. Function bodies are
byte-identical to origin/master; the new file imports net and strings
(the union the moved declarations use), which drop out of the base
import block.

Validation: gofmt -l clean, go build ./pkg/config/ green.
